🐛 Fix extra support for enum with its own schema
